🏗 use lit for hal
It's quite similar, in all honesty. In my estimation, it makes certain
aspects more "rigorous" or "normal", at the cost of being a little
mystifying (on the other hand, regular web components are mystifying!).

I expect the same operation, done on the agent element, to be similarly
minor. I think most of what lit is simplifying is the view code for a
custom element, which just isn't all that relevant for these two!