Doc/add hoop standalone doc (#8)
Nearly a no-op change on the library (just started exporting the AgentElement class, and renamed orb-state -> status).

Primarily this is a refactor to the index page to add a standalone hoop element, and provide some very bare-bones documentation on how to use it directly. Also exposes a bit of the audio functions to make it easier to work with the animation directly.